refactor(person): Reorganize person module tests and services
This commit refactors the person module by reorganizing the tests and services into a more modular structure. The monolithic test files have been broken down into smaller, more focused files under `tests/suites/person/` and `tests/person/`.

Components affected:
- `banking-db-postgres`: Moved person repository tests to a dedicated `person` module.
- `banking-logic`: Moved person service tests to a dedicated `person` module.
- `banking-api`: Minor updates to error handling and service definitions.
- `docs`: Updated testing and error handling guidelines.

Benefits:
- Improved test organization and maintainability.
- Clearer separation of concerns within the person module.
- Enhanced developer experience when working with person-related components.
