+++ banking-api/src/error.rs
@@ -14,29 +14,31 @@ pub enum BankingError {
     NotFound(String),
     #[error("Person service error: {0}")]
     PersonServiceError(#[from] crate::service::PersonServiceError),
+    #[error("Country service error: {0}")]
+    CountryServiceError(#[from] crate::service::CountryServiceError),
     #[error("Audit service error: {0}")]
     AuditServiceError(#[from] crate::service::AuditServiceError),
     // Account-related errors
     #[error("Account not found: {0}")]
     AccountNotFound(Uuid),
-    
+
     #[error("Account {account_id} is frozen: {frozen_reason}")]
-    AccountFrozen { 
-        account_id: Uuid, 
-        frozen_reason: String 
+    AccountFrozen {
+        account_id: Uuid,
+        frozen_reason: String,
     },
-    
+
     #[error("Account {account_id} was closed on {closure_date}")]
-    AccountClosed { 
-        account_id: Uuid, 
-        closure_date: NaiveDate 
+    AccountClosed {
+        account_id: Uuid,
+        closure_date: NaiveDate,
     },
-    
+
     #[error("Insufficient funds in account {account_id}: requested {requested}, available {available}")]
-    InsufficientFunds { 
-        account_id: Uuid, 
-        requested: Decimal, 
-        available: Decimal 
+    InsufficientFunds {
+        account_id: Uuid,
+        requested: Decimal,
+        available: Decimal,
     },
 
     #[error("Account {account_id} is not operational: {reason}")]
@@ -46,63 +48,61 @@ pub enum BankingError {
     },
 
     #[error("Account {account_id} is not in a transactional state")]
-    AccountNotTransactional {
-        account_id: Uuid,
-    },
-    
+    AccountNotTransactional { account_id: Uuid },
+
     // Customer-related errors
     #[error("Customer not found: {0}")]
     CustomerNotFound(Uuid),
-    
+
     #[error("Customer {customer_id} is deceased (date of death: {date_of_death})")]
-    CustomerDeceased { 
-        customer_id: Uuid, 
-        date_of_death: NaiveDate 
+    CustomerDeceased {
+        customer_id: Uuid,
+        date_of_death: NaiveDate,
     },
-    
+
     #[error("Customer {customer_id} is blacklisted: {blacklist_reason}")]
-    CustomerBlacklisted { 
-        customer_id: Uuid, 
-        blacklist_reason: String 
+    CustomerBlacklisted {
+        customer_id: Uuid,
+        blacklist_reason: String,
     },
-    
+
     // Transaction-related errors
     #[error("Transaction limit exceeded: attempted {attempted}, limit {limit} for {limit_type:?}")]
-    TransactionLimitExceeded { 
-        limit: Decimal, 
-        attempted: Decimal, 
-        limit_type: LimitType 
+    TransactionLimitExceeded {
+        limit: Decimal,
+        attempted: Decimal,
+        limit_type: LimitType,
     },
-    
+
     #[error("Invalid signature: required {required_signatories:?}, provided {provided_signatories:?}")]
-    InvalidSignature { 
-        required_signatories: Vec<Uuid>, 
-        provided_signatories: Vec<Uuid> 
+    InvalidSignature {
+        required_signatories: Vec<Uuid>,
+        provided_signatories: Vec<Uuid>,
     },
-    
+
     #[error("Approval required for transaction {transaction_id}: required approvers {required_approvers:?}")]
-    ApprovalRequired { 
-        transaction_id: Uuid, 
-        required_approvers: Vec<Uuid> 
+    ApprovalRequired {
+        transaction_id: Uuid,
+        required_approvers: Vec<Uuid>,
     },
-    
+
     // Compliance-related errors
     #[error("Compliance violation: {violation_type} for customer {customer_id:?}")]
-    ComplianceViolation { 
-        violation_type: String, 
-        customer_id: Option<Uuid> 
+    ComplianceViolation {
+        violation_type: String,
+        customer_id: Option<Uuid>,
     },
-    
+
     #[error("KYC incomplete for customer {customer_id}: missing documents {missing_documents:?}")]
-    KycIncomplete { 
-        customer_id: Uuid, 
-        missing_documents: Vec<String> 
+    KycIncomplete {
+        customer_id: Uuid,
+        missing_documents: Vec<String>,
     },
-    
+
     #[error("Sanctions match for customer {customer_id}: {match_details}")]
-    SanctionsMatch { 
-        customer_id: Uuid, 
-        match_details: String 
+    SanctionsMatch {
+        customer_id: Uuid,
+        match_details: String,
     },
 
     // Agent Network Hierarchy Limit Violations
@@ -134,45 +134,43 @@ pub enum BankingError {
 
     // Calendar and Business Day Validation
     #[error("Invalid weekend days configuration: {invalid_days:?} - days must be between 1 (Monday) and 7 (Sunday)")]
-    InvalidWeekendDays {
-        invalid_days: Vec<i32>,
-    },
+    InvalidWeekendDays { invalid_days: Vec<i32> },
 
     #[error("Weekend configuration validation failed: {validation_errors:?}")]
     WeekendConfigValidationFailed {
         validation_errors: Vec<String>,
     },
-    
+
     // Product and system errors
     #[error("Invalid product id: {0}")]
     InvalidProductId(Uuid),
 
     #[error("Product not found: {0}")]
     ProductNotFound(Uuid),
-    
+
     #[error("Product catalog unavailable for {product_id}, fallback used: {fallback_used}")]
     ProductCatalogUnavailable {
         product_id: Uuid,
-        fallback_used: bool
+        fallback_used: bool,
     },
-    
+
     #[error("Business day calculation error for date {date} in jurisdiction {jurisdiction}")]
-    BusinessDayCalculationError { 
-        date: NaiveDate, 
-        jurisdiction: String 
+    BusinessDayCalculationError {
+        date: NaiveDate,
+        jurisdiction: String,
     },
-    
+
     // Network and infrastructure
     #[error("Network error: {error_details}, retry possible: {retry_possible}")]
-    NetworkError { 
-        error_details: String, 
-        retry_possible: bool 
+    NetworkError {
+        error_details: String,
+        retry_possible: bool,
     },
-    
+
     #[error("Database constraint violation: {constraint} - {details}")]
-    DatabaseConstraintViolation { 
-        constraint: String, 
-        details: String 
+    DatabaseConstraintViolation {
+        constraint: String,
+        details: String,
     },
 
     // Validation errors
@@ -188,24 +186,24 @@ pub enum BankingError {
         field: String,
     },
 
-    // Transaction errors  
+    // Transaction errors
     #[error("Invalid transaction amount: {0}")]
     InvalidTransactionAmount(String),
-    
+
     #[error("Transaction not found: {0}")]
     TransactionNotFound(String),
-    
+
     #[error("Validation failed: {0}")]
     ValidationFailed(String),
-    
+
     // Date/Time errors
     #[error("Date calculation error: {0}")]
     DateCalculationError(String),
-    
+
     // Document errors
     #[error("Duplicate identity document: {0}")]
     DuplicateIdentityDocument(String),
-    
+
     // Authorization errors
     #[error("Unauthorized operation: {0}")]
     UnauthorizedOperation(String),
@@ -213,11 +211,10 @@ pub enum BankingError {
     // Internal errors
     #[error("Internal error: {0}")]
     Internal(String),
-    
+
     // Not implemented features
     #[error("Feature not implemented: {0}")]
     NotImplemented(String),
-
 }
 
 impl From<anyhow::Error> for BankingError {
@@ -240,7 +237,7 @@ impl From<sqlx::Error> for BankingError {
                 } else {
                     BankingError::Internal(format!("Database error: {}", db_err.message()))
                 }
-            },
+            }
             _ => BankingError::Internal(format!("Database error: {err}")),
         }
     }
@@ -251,14 +248,12 @@ impl From<Box<dyn std::error::Error + Send + Sync>> for BankingError {
     }
 }
 
-
-
 #[derive(Debug, Clone, Serialize, Deserialize)]
-pub enum LimitType { 
-    Daily, 
-    Monthly, 
-    PerTransaction, 
-    Terminal, 
-    Branch, 
-    Network 
+pub enum LimitType {
+    Daily,
+    Monthly,
+    PerTransaction,
+    Terminal,
+    Branch,
+    Network,
 }
