+++ banking-db-postgres/tests/suites/person/mod.rs
@@ -1,0 +1,7 @@
+pub mod messaging_repository_tests;
+pub mod location_repository_tests;
+pub mod locality_repository_tests;
+pub mod country_subdivision_repository_tests;
+pub mod country_repository_tests;
+pub mod person_repository_tests;
+pub mod helpers;
