+++ banking-logic/tests/person/mod.rs
@@ -1,0 +1,15 @@
+pub mod common;
+pub mod country_tests;
+pub mod mock_country_repository;
+pub mod country_subdivision_tests;
+pub mod mock_country_subdivision_repository;
+pub mod locality_tests;
+pub mod mock_locality_repository;
+pub mod location_tests;
+pub mod mock_location_repository;
+pub mod messaging_tests;
+pub mod mock_messaging_repository;
+pub mod entity_reference_tests;
+pub mod mock_entity_reference_repository;
+pub mod person_tests;
+pub mod mock_person_repository;
